refactor(nat): Do not pass around next_header when unnecessary
Not sure if passing around the next_header type around between functions
for the stateful NAT logic was ever used; now at least, it's not. Let's
remove it.
